+++ src/index.pug
@@ -14,7 +14,7 @@ html(lang="ja")
         | Creators' Conferenceは、4/14(火)に開催予定だった複数の創作系サークルが合同で行うサークル説明会です。
         | 新型コロナウイルスの感染拡大の影響で、大学側よりサークル活動の自粛が要請されたため、中止になりました。
         | 本サイトでは、各サークルの活動団体別の発表で使用される予定だったスライドなどを掲載しています。
-      div#categories
+      #categories
         each category in categories
           +category(category)
     include includes/footer.pug
