+++ src/js/details.js
@@ -1,0 +1,11 @@
+export const setupDetails = () => {
+  const $$button = document.querySelectorAll(".toggle-details")
+  for (const $button of $$button) {
+    const $details =
+      $button.parentElement.querySelector(".details") ??
+      $button.closest(".details")
+    $button.addEventListener("click", () => {
+      $details.classList.toggle("hidden")
+    })
+  }
+}
