+++ src/mixins/category.pug
@@ -1,8 +1,8 @@
 include ./group.pug
 
 mixin category(category)
-  div.category
+  .category
     h2.name= category.name
-    div.groups
+    .groups
       each group in category.groups
         +group(group)
